Separate mutation and muration state.
